Move task state dataclasses to their own module
